[#20] Split validation functions out to separate validator module
